+++ lib/rspec/core/formatters/html_snippet_extractor.rb
@@ -24,6 +24,8 @@ module RSpec
         @@converter = NullConverter
         begin
           require 'coderay'
+          RSpec::Support.require_rspec_core 'source/syntax_highlighter'
+          RSpec::Core::Source::SyntaxHighlighter.attempt_to_add_rspec_terms_to_coderay_keywords
           @@converter = CoderayConverter
           # rubocop:disable Lint/HandleExceptions
         rescue LoadError